refactor(centralized-paper): move to components folder
