+++ src/components/CentralizedPaper.js
