+++ src/components/ReportTable/ReportTable.test.js
@@ -2,7 +2,7 @@ import Adapter from 'enzyme-adapter-react-16'
 import Enzyme from 'enzyme'
 import React from 'react'
 import { shallow, mount, render } from 'enzyme'
-import CentralizedPaper from './CentralizedPaper'
+import CentralizedPaper from '../CentralizedPaper'
 import Paper from '@material-ui/core/Paper'
 import ReportTable from './ReportTable'
 import Table from '@material-ui/core/Table'
